Agent-based installation late binding
This enhancement describes a change to agent-based installation whereby
agents are not necessarily bound to cluster definitions, and that
binding may be done dynamically at a later time.  This is different from
today where a Discovery ISO is associated with a cluster definition and
all agents that come up from that ISO are also bound to that cluster
definition.